celery_client_and_worker/requirements.txt
celery>=5.4.0
chardet>=5.2.0
colorama>=0.4.6
Flask-Bootstrap>=3.3.7.1
Flask-Mail>=0.10.0
pillow>=10.3.0
PyPDF2>=3.0.1
pymongo>=4.7.3
python-dotenv>=1.0.1
redis>=5.0.6
reportlab>=4.2.2
